unify_auth - pull BaseDns back into dns_provider, both considered legacy. Removed BaseHttp; implement chal_type; fixup client and tests to match
