+++ sewer/auth.py
@@ -4,16 +4,27 @@ from typing import Sequence, Tuple
 from hashlib import sha256
 from sewer.lib import safe_base64
 
-### FIX ME ### a few existing providers assume self.log_response exists... need to convert 'em
 
-from sewer.lib import log_response as lib_lr
+class ValidationMethod:
+
+    def __init__(self, label: str, id_type: str, supported: bool):
+        self.label = label
+        self.id_type = id_type
+
+supported_validations = {
+    "http-01": ValidationMethod("http-01", "dns", True),
+    "dns-01": ValidationMethod("dns-01", "dns", True),
+}
 
 
 class BaseAuth(object):
-    """
-    Abstract base class for providers of key authorizations that are responsible
-    for taking the challenge info they're given and posting it where the ACME server
-    can find it, as well as for cleaning up after validation is complete.
+    """ Abstract base class for providers of key authorizations that are
+    responsible for taking the challenge info they're given and posting it
+    where the ACME server can find it, optionally verifying that the
+   challenge is properly published, as well as for cleaning up after
+    validation is complete.
+
+    NB: BaseDns is a shim for legacy DNS providers, and BaseHttp is deprecated.
 
     There are layers of interface for the challenge setup & clear operations.  The
     newest, and the one that will be called directly from sewer's ACME code, passes
@@ -48,13 +59,21 @@ class BaseAuth(object):
 
     def __init__(self, **kwargs):
         """
-        __init__ accepts logging control argument(s).  From preferred to legacy:
+        chal_types      Sequence[str] set of implemented ACME challenge types
+
+        optional logging control argument(s).  From preferred to legacy:
          *  logger	a configured logging object to use as-is
          *  log_level	level to setup local logger for, defaulting to INFO
          *  LOG_LEVEL   obsolescent legacy version of log_level
 
         logger will override log_level will override LOG_LEVEL if more than one is passed
         """
+
+        ### FIX ME ### auth_type handling?  and that's not best named, should be challenge_type
+
+        self.chal_types = kwargs.pop("chal_types")
+        if not all(ct in supported_validations for ct in self.chal_types):
+            raise ValueError("Unsupported challenge type in list: %s" % self.chal_types)
 
         logger = kwargs.pop("logger", None)
         log_level = kwargs.pop("LOG_LEVEL", "INFO")	# compatibility / obsolescent
@@ -72,9 +91,6 @@ class BaseAuth(object):
                 formatter = logging.Formatter("%(message)s")
                 handler.setFormatter(formatter)
                 self.logger.addHandler(handler)
-
-        ### hack for compatibility with some (3?) existing dns providers
-        self.log_response = lib_lr
 
     def setup_cert(self, authorizations: Sequence[Tuple[str, str, str]]):
         for auth in authorizations:
@@ -100,93 +116,3 @@ class BaseAuth(object):
     def clear_auth(self, domain: str, token: str, key_auth: str):
         "clear one authorization - remove dns record, file, etc."
         raise NotImplementedError("clear_auth method must be implemented when clear_cert is not.")
-
-
-def dns_challenge(key_auth: str) -> str:
-    "return safe-base64 of hash of key_auth; used for dns response"
-
-    return safe_base64(sha256(key_auth.encode("utf8")).digest())
-
-
-class BaseDns(BaseAuth):
-    """
-    BaseDns provides a compatibility layer in the *_auth methods.  These convert
-    the "standard" three arguments into the two that are needed by legacy dns
-    providers, allowing them to work behind these new interfaces.  That's not to
-    say that they have to do that, and if a provider can update multiple texts at
-    a time, it can take advantage of the *_cert interfaces.  Whatever works best...
-    """
-
-    def __init__(self, **kwargs):
-        super(BaseDns, self).__init__(**kwargs)
-        self.auth_type = "dns-01"
-
-    def setup_auth(self, domain: str, token: str, key_auth: str):
-        "shim to use legacy create_dns_record method"
-
-        self.create_dns_record(domain, dns_challenge(key_auth))
-
-    def clear_auth(self, domain: str, token: str, key_auth: str):
-        "shim to use legacy delete_dns_record method"
-
-        self.delete_dns_record(domain, dns_challenge(key_auth))
-
-    ### FIX ME ### left the legacy DNS provider methods here for compatibility and documentation
-
-    def create_dns_record(self, domain_name, domain_dns_value):
-        """
-        Method that creates/adds a dns TXT record for a domain/subdomain name on
-        a chosen DNS provider.
-
-        :param domain_name: :string: The domain/subdomain name whose dns record ought to be
-            created/added on a chosen DNS provider.
-        :param domain_dns_value: :string: The value/content of the TXT record that will be
-            created/added for the given domain/subdomain
-
-        This method should return None
-
-        Basic Usage:
-            If the value of the `domain_name` variable is example.com and the value of
-            `domain_dns_value` is HAJA_4MkowIFByHhFaP8u035skaM91lTKplKld
-            Then, your implementation of this method ought to create a DNS TXT record
-            whose name is '_acme-challenge' + '.' + domain_name + '.' (ie: _acme-challenge.example.com. )
-            and whose value/content is HAJA_4MkowIFByHhFaP8u035skaM91lTKplKld
-
-            Using a dns client like dig(https://linux.die.net/man/1/dig) to do a dns lookup should result
-            in something like:
-                dig TXT _acme-challenge.example.com
-                ...
-                ;; ANSWER SECTION:
-                _acme-challenge.example.com. 120 IN TXT "HAJA_4MkowIFByHhFaP8u035skaM91lTKplKld"
-                _acme-challenge.singularity.brandur.org. 120 IN TXT "9C0DqKC_4MkowIFByHhFaP8u0Zv4z7Wz2IHM91lTKec"
-            Optionally, you may also use an online dns client like: https://toolbox.googleapps.com/apps/dig/#TXT/
-
-            Please consult your dns provider on how/format of their DNS TXT records.
-            You may also want to consult the cloudflare DNS implementation that is found in this repository.
-        """
-        raise NotImplementedError("create_dns_record method must be implemented.")
-
-    def delete_dns_record(self, domain_name, domain_dns_value):
-        """
-        Method that deletes/removes a dns TXT record for a domain/subdomain name on
-        a chosen DNS provider.
-
-        :param domain_name: :string: The domain/subdomain name whose dns record ought to be
-            deleted/removed on a chosen DNS provider.
-        :param domain_dns_value: :string: The value/content of the TXT record that will be
-            deleted/removed for the given domain/subdomain
-
-        This method should return None
-        """
-        raise NotImplementedError("delete_dns_record method must be implemented.")
-
-
-class BaseHttp(BaseAuth):
-    """
-    Since support for http providers is new at about the same time as the new
-    interface, there's no need for backwards compatible shims in BaseHttp.
-    """
-
-    def __init__(self, **kwargs):
-        super(BaseHttp, self).__init__(**kwargs)
-        self.auth_type = "http-01"
