--- sewer/http_providers/__init__.py
@@ -1,1 +1,0 @@
-from .common import BaseHttp  # noqa: F401
