--- sewer/http_providers/common.py
@@ -1,1 +1,0 @@
-from sewer.auth import BaseHttp
